notifier/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-restx==1.3.0
Flask_JWT_Extended==4.6.0
Flask-SocketIO==5.4.1
requests==2.32.2
pyyaml==6.0.2
eventlet==0.37.0
Werkzeug==3.1.1